Assure NORMAL flow present on ovs_bridges (bsc#1097252)
There are some cases where the host system could end up in a state
which causes NORMAL flows to not be present in bridges which are used
for management or other, external host access. This would result in
complete loss of network connection to host.

One possible case for a SLES based node is when the host is shutdown
cold by turning poweroff  immediately and then rebooted.

Since all ovs bridges configured during bootup are meant to provide
network connection to the host, we can always assure we add the NORMAL
flow to those switches.

Change-Id: I7f4aa5942950db349ee6b3133f24b921fe502d02