Hugepage directories need a valid user (bsc#1092431)
When creating a directory for hugepages, it is possible
that the qemu user has not yet been created.
That would cause a failure in execution.

the "root" user should already be created.

It is nova (qemu binaries) that need access to the directory. 
So we have to wait until nova is installed to change the
directory ownership. 
see https://gerrit.suse.provo.cloud/#/c/4456

Change-Id: I3d5a0dfd4977fd1b85140e34d3eca0eb0cbf198c