Restart dpdk network after restarting openvswitch (bsc#1104407)
If an ip address is assigned to an openvswitch dpdk
bridge, that bridge needs to be brought up by using
"systemctl restart network"  any time openvswitch itself
is restarted. Otherwise any ip address associated with
the bridge will not be configured on the system.

In the case of vxlan, this could result on the
"local_ip" not being found on the host, causing
neutron-openvswitch-agent to fail start.

Change-Id: I3b6111498c98b31ca07eed5f798f8c489fed4542
(cherry picked from commit 40f1f9f2f15e8445925702fade32c59978f8067c)