SCRD-5116 Support root LV named root or LVRoot
Switching to using kiwi built images means that the LV containing the
root file system will be named LVRoot as that is the hardcoded name
that kiwi uses, which we can't override.

As a result the osconfig-ansible diskconfig role needs some updating
to dynamically determine the name of the root LV. As part of this we
are removing some legacy code left over from the rebranding efforts,
and also making the code dynamically determine the root VG name also.
Additionally some of the logic has been simplified by using alternate
commands to determine and retrieve the information, as well as avoid
the need for ansible loop primitives.

Reformat some actions using the preferred multi-line key: value style
rather than the key=value style for parameters.

Corrected the name of the ansible role in the name: statements to be
consistent.

Change-Id: I904db17c59731685c36bea1459c2cd91ffc2fe57
(cherry picked from commit f41f8e501ca2ec50258cc5dc74e35db6c3e87f17)