Do not rewrite a manually written multipath.conf file (SCRD-7840)
To allow the for a node to boot from SANs in which the administrator
manually crafted a /etc/multipath.conf, osconfig playbooks will not
be overwriting the file. To do this, the set a the variable

   manual_multipath_conf

to True in [1].

Also fixed the template file [2] so that any parameter values under
multipath_defaults will not cause site.yml to fail with jinja2 errors.

[1] ~/openstack/my_cloud/config/multipath/multipath/multipath_settings.yml
[2] roles/multipath/templates/multipath.conf.j2

Change-Id: Ic37d33944a10049de21d9e63b6c1c9386e362492