Remove OVS reboot workaround (SCRD-7979)
With the deprecation of external_network_bridge and the introduction
of the native OVSDB/OpenFlow interface in the neutron OVS agent, it
is no longer safe to apply forward NORMAL rules outside of neutron.
Under the right conditions, this practice introduces an L2 loop into
the underlay network. The reason for the workaround was to enable
management network access over neutron-controlled OVS bridges. For
loop pruning and other reasons such as security and other architectural
best practices, it is not recommended to allow management network access
over OVS bridges under the control of neutron. Removing these reboot
workarounds protects the underlay from L2 loops.

Change-Id: I968400aa5ba48d561cd5afeb7c720c53a3bf9b49