SCRD-8041 Ensure neutron-server running after network restart (SCRD-8041)
In some cases the restart of the network after reconfiguring a network
interface can lead to neutron-server being stopped, and it doesn't get
automatically restarted.

Add a workaround for this in network_interface-reconfigure.yml by adding
a task to ensure it is restarted after a network restart, similarly to
what is done for the neutron-openvswitch-agent service.

Change-Id: I9e6978d9895088723a12f22eebc907d22431d499