Bump fs.inotify.max_user_instances (SCRD-8529)
It turns out that on environment with a lot of network resources
the default value of fs.inotify.max_user_instances is not
enough and the dnsmasq process spawned by the neutron-dhcp-agent
hits the limit.

This patch bumps the fs.inotify.max_user_instances parameter
from 128 to 256 in all nodes that have the neutron-dhcp-agent
installed.

Change-Id: I1d2cc8f8f0103d2b499ee4ceedece88952473b7c
Signed-off-by: Antonio Ojea <aojeagarcia@suse.com>