Ensure ardana-update-pkgs works for dist-upgrade (SOC-9857)
Issues: SOC-9857, SOC-9844

SOC-9857
========
When running during an upgrade we will need to switch the set of repos
we have configured from Cloud8/SLE12SP3 to Cloud9/SLE12SP4 so that the
zypper dist-upgrade will be able to upgrade the base system.

Remove override of suse_release setting in check-repo-status.yml; for
the upgrade we want to use the value configured in the product-media
role.

SOC-9844
========
Ensure we remove the stale Cloud8/SLE12SP3 PTF repo and add the new
Cloud9/SLE12SP4 PTF repo.

Depends-On: I63dd3c14344521b360b87ad7a56638501403372a
Change-Id: Ic8d6c466c91a2b2771788cd81787faf34feb6a74