Add an global_filter entry to lvm.conf  (bsc#1140512)
pvscan does not use the lvm.conf filter but instead the
global_filter. This changes adds the global_filter to address
case where pvscan incorrectly grabs a multipath cinder volume
that contains an LVM2 signature. When this occurs the host locks
the cinder volume incorrectly. With a global_filter this "grab"
and "lock" do not occur.

Change-Id: Ie92e04b4244923f294087407360c13c50f36527c