Introduce conditional forward:NORMAL rule on POST-UP for OVS bridges (SOC-9939)
This change introduces a conditional insertion of a forward:NORMAL
openflow rule for OVS bridges. If forward_normal_on_post_up is true
as set by the configuration processor, the Wicked POST-UP actions for
the device will contain a step that ensures the bridge is forwarding
traffic by inserting an openflow rule that says "act like a normal
bridge".

Change-Id: Ie19b7cbfa0ba23310da4f3e006af3b18ae4dbcad