Ensure ovs_user and ovs_group defined (SOC-11149)
In the default openvswitch configuration used in SLE 12 SP3/SOC 8 the
user and group used to manage the openvswitch services' log file are
'root' and 'root', and don't appear to be specified via any config
file settings.

However in a SLE 12 SP4/SOC 9 fresh deployment the user and group
values are specified via the OVS_USER_ID, which encodes the user and
group as a colon (:) separated pair, potentially in two places:
  * /etc/openvswitch/default.conf
    Specifies OVS_USER_ID as "root:root" for backwards compatibility
    reasons.
  * /etc/sysconfig/openvswitch
    Will include an OVS_USER_ID setting that specifies the user and
    group as "openvswitch:hugetlbfs", but only on a fresh install;
    this entry will not exist on a migrated/dist-upgraded system.

When upgrading from SOC 8 to SOC 9 there won't be any OVS_USER_ID
entry in /etc/sysconfig/openvswitch, so we need to ensure we fallback
to using the backwards compatibility settings as specified in the
/etc/openvswitch/default.conf file, which will also match those values
used by the running openvswitch services.

This patch updates the process by which the ovs_user and ovs_group
values are determined to line up with the above hierarchy, first
checking for an OVS_USER_ID setting in /etc/sysconfig/openvswitch
and then falling back to check in /etc/openvswitch/default.conf or
failing that uses a default value of "root:root".

The patch also optimises the process somewhat such that we only
lookup the OVS_USER_ID setting in the relevant file once, and then
extract the user and group values out of the result, rather than
repeating the process for each value.

Change-Id: I9e1f264373170a8e9f30044ea35e8845ebbbb75d